Move custom widgets to `src/gui`.
Also, bugfix: `chkbox_state` -> `checkbox_state`.
